refactor(codegen): extract std::decimal::* into stdlib/decimal.rs (Round 1.4 PoC)
Single-fn proof of concept for the codegen model-org refactor. Lifts
lower_std_decimal_to_float into stdlib/decimal.rs as a pub(crate)
DecimalStdlib<'ctx> trait extension impl'd on Cx<'ctx, 'p>. Trait
imported at the top of codegen.rs so the two existing dispatch sites
(statement + expression position) keep the self.lower_std_*(...) call
shape with no change.

Surface bumps to pub(crate):
- struct Cx<'ctx, 'p> + fields context, builder, module
- enum CodegenTy
- struct Scope<'ctx>
- fn Cx::lower_expr

No public API change. Pattern validated; subsequent sub-rounds scale
the same shape across the other ~24 std::* namespaces.
